Add latency threshold to mcrouter destinations
Summary:
In an attempt to address the 'bad box problem', Andrii, David, and I discussed last night and decided on a solution that bases TKO decisions on average latency. This is a basic implementation. The following goals were kept in mind for a good solution:

1) Simple: this adds less than 10 lines of code to ProxyDestination
2) Robust to outliers: This puts a cap on requests with huge latency, ensuring that at least 2 such requests are required to cause a TKO. In addition, since each proxy maintains independent state, if one proxy thread suffers from a small number of outliers while the others remain fine, the others will continually reset the failure count.
3) Easy to understand and configurable: At its core, this just defines an average latency we don't want a destination to go above.
4) Safe: In conjunction with tko limits we can test this without risking too many boxes going TKO.
5) Progressive: The worse a box is, the more we want to avoid it. In this case, since each webi maintains an independent view of the world, we expect each one to see a different average latency in some distribution. If a box is just sort of bad, only some webis will TKO him, allowing for a nice removable of load. We also have the benefit of feedback. As we reduce load and he gets faster, mcrouters will start unmarking him.

One downside of this solution is that because it doesn't include the latency of probes (should we?), webis will keep having to revisit hosts every so often (probe-timeout-initial determines this length, currently 10 seconds) even if they remain very slow. It can also take a while to flush the window, and we may TKO a host several times even after he goes back to normal (potential solution: weighted moving average)

Test Plan: Integration tests. Unit test for moving average

Reviewed By: andrii@fb.com

Subscribers: markdrayton, sameen

FB internal diff: D1410264

Tasks: 4589734