Store request logger and request context in the same allocation
Summary:
This shows a perf improvement, since we remove an allocation on the fast path.

1. Instead of a ProxyRequestContext holding a unique_ptr to
a ProxyRequestLogger we create a new wrapper, LoggingProxyRequestContext (which is a point of customization)
which now contains both a ProxyRequestContext and a ProxyRequestLogger (plus whatever other logging is desired).

2. Another change is moving ProxyMcRequest into ProxyMcRequest.h/.cpp, which is a simple refactoring.

3. Some dead code cleanup

Test Plan:
unit tests, mchammer benchmark

Reviewed By: pavlo@fb.com

Subscribers: trunkagent, meisner

FB internal diff: D1451725

Tasks: 4745545