Simplify proxy config lock; remove ThreadReadLock
Summary:
Currently config_lock is only needed to protect swapping of proxy->config pointer.
Historically the lock was much less fine-grained, and it was diffcult to tell when it was required to hold it.
However now we can encode the intention directly. This diff makes the lock and the pointer private and
introduces two thread-safe functions:

- getConfig(): returns a shared_ptr to a valid config. The config stored in proxy might change during the
lifetime of a single request, but since we save the shared_ptr to the original config in the context,
we will route using the original config only
- swapConfig(): atomically swap the shared_ptr to a new value; return the previous value

All other uses of config_lock are clearly unnecessary and can be removed, reducing deadlock risk and improving
performance.

In addition, I also removed config_ pointer from proxy_request_t as it was unused. The intention is
to eventually get rid of proxy_request_t completely since ProxyRequestContext plays the same role.

Test Plan:
unit tests pass
--continuously-reload-configs

Facebook:
Will canary on a few fabric boxes

Reviewed By: stepan@fb.com

FB internal diff: D1480674