Zero-copy for umbrella
Summary:
title
This alone is not a zero-copy yet, will need to rebase on top of all other zero-copy diffs before we will be able to test performance of this.
But it should be ok for review, since it doesn't conflict with other parts.

Test Plan: tests

Reviewed By: brianp@fb.com

Subscribers: trunkagent, alikhtarov

FB internal diff: D1538327

Tasks: 4889426