Make mc_res_connect_timeout a hard error
Summary: If we end up shortening mc_res_connect_timeout from the max request time mc_res_connect_timeout should be soft so that retransmits don't hard TKO a box.
hort of that, if we can't even establish a connection in the time allocated for a request, the server is likely dead.

Test Plan: unit tests

Reviewed By: pavlo@fb.com

Subscribers: trunkagent, sameen, ipozgaj, alikhtarov

FB internal diff: D1506290