Only unmark TKO on probes
Summary: We don't want outstanding requests to unmakr TKO on a box. This creates an unpredictable scenario for TKO, as instead of needing 3 consecutive failures or whatever, we actually may need 20 because of outstanding requests. It also undermines initial probe delay and the predictability of that

Test Plan: unit tests

Reviewed By: pavlo@fb.com

Subscribers: meisner, trunkagent, sameen, alikhtarov

FB internal diff: D1587681

Tasks: 5275894