onRequest API changes
Summary:
This is an API change; functional changes have been kept to the necessary minimum.

There are two interface changes in this diff:

- McServerRequestContext is now handed off to the user (via an rvalue reference),
and the user must surrender it back via a static reply() method.
From API design point of view, this ensures that reply is called exactly once
(context's destructor will assert that reply was called, and the context itself
can only be moved around).  It also allows an optimization:
we don't need any heap allocated storage for the context on the server side;
any context storage for async requests is up to the user, and allows for stack only storage
in cases where we can return the reply immediately.

- McRequest is also passed via an rvalue reference. It's up to the user to move it somewhere if async call is necessary. Same performance rationale as above.

Test Plan: unit tests

Reviewed By: andrii@fb.com

Subscribers: trunkagent, alikhtarov

FB internal diff: D1613935

Tasks: 5372756