Simplify asynclog logic + code cleanup
Summary:
This removes complicated async logic. The end result we want is to wait until the asynlog is written to disk on the awriter thread
before returning the reply. We can easily accomplish this with a baton, and remove a lot of old code.

Also removes dead async writer code that's not used anymore (we write files to disk from stats thread directly).

Test Plan: unit tests

Reviewed By: pavlo@fb.com

Subscribers: trunkagent, alikhtarov

FB internal diff: D1752910