Simplify ProxyDestinationMap
Summary: this diff fixes one more bug: when we call removeAllUnused, we assume the old config is already destroyed. But if there is some pending request, the config won't be destroyed.
So removeAllUnused does not remove destinations that were removed from old config *and* had pending request. These destinations still exist, thus, ProxyClientShared is not destroyed as well.
If the same "host:port:timeout" is added next time to config, we will reuse ProxyClientShared with broken TkoTracker.

Solution:
There is no reason for ProxyDestinationMap to own ProxyDestinations (i.e. use shared_ptr). It may simply have weak_ptrs. In this case once the old config is destroyed, all the ProxyDestinations not used in the new one will be destroyed as well.

Test Plan: mcrouter unit tests

Reviewed By: @alikhtarov

Differential Revision: D1789044