Add option to group remote errors.
Summary: Add a new option (--group-remote-errors) to return mc_res_remote_error
for all remote errors returned by mcrouter.

The main goal of this change is to stop marking underlying mcrouter as
TKO if it is still alive. Whenever we have a communication like this:
client<->mcrouter1<->mcrouter2<->memcached and the memcached stops
responding, we should not mark mcrouter2 as TKO. This patch solves this
issue.

Test Plan: Unit tests and manual tests.

Reviewed By: @pavlo-fb

Differential Revision: D1856822