Reduce TimeoutController bucket explosion
Summary: The way we use timeouts when fibers are enabled causes the TimeoutController to create tens of thousands of buckets, since the wait duration is calculated in nanoseconds. The buckets are iterated linearly each time we do a timed wait. This diff bypasses the problem by converting the timeout to milliseconds - the set of possible buckets should be drastically reduced as a result.

Test Plan: unit tests

Reviewed By: @andriigrynenko

Differential Revision: D1872170