Implement default exception callback for FiberManager
Summary: Default implementation outputs information about the exception and context in which exception was caught. After that exception is just rethrown.
This strictly improves the behavior, because previously any call into exceptionCallback_ would result in bad_function_call because it was never set.

Test Plan: Added some throws in different places in mcrouter code and checked that they are reported correctly

Reviewed By: @pavlo-fb

Differential Revision: D1871588