Reconnect on probe if connection was closed
Summary: we mark destination as active on every send. So it is possible that between two sends we reset client in ProxyDestination. If probe delay is > inactive period, we'll reset the connection between consecutive probes. It is still fine, as lonbg as we reconnect when we send a probe. For example, if probe_max_delay is 10 minutes, we don't want to keep the connection opened for hours if the box is unresponsive. Just reconnect and send probe.

Test Plan: mcrouter unit tests, unit test attached

Reviewed By: @alikhtarov

Differential Revision: D1920577