Replacing FBI_ASSERTS in mcrouter with PCHECKs and asserts
Summary: PCHECK(expr) was used for cases in which the process will definitely fail and won't be able to continue.
For all other cases, assert(expr) was used.

For files under /mcrouter/lib/mc and /mcrouter/lib/fbi, assert(expr) was used since that is legacy C code and cannot be linked to glog's PCHECK.

Test Plan: fbmake runtests

Reviewed By: @alikhtarov

Differential Revision: D1913685