Move LoggingRoute on top of McrouterHandleIf
Summary: In a future diff, we'll need access to ProxyRequestContext and McrouterInstance. Previously it wasn't possible to do this as LoggingRoute was under lib/ and based on the abstract `RouteHandleIf`.

This diff moves the LoggingRouter out of lib/, removes the `RouteHandleIf` templating and builds it on top of McrouterHandleIf.

Test Plan: - Unit tests
 - Confirmed logs are still emitted

Reviewed By: @pavlo-fb

Differential Revision: D2063115