In the standalone use case, add ip address to ProxyRequextContext
Summary: LoggingRoute can now access the IP of the machine sending the request to us through ProxyRequestContext.

To achieve this, adding logic to McServerSession to pass this data around the call chain eventually to McRouterClient, which will then embed the data to ProxyRequestContext dependent on a setting. The setting can be enabled using the `--retain-source-ip` command line flag.

Test Plan: - Unit tests

Reviewed By: @alikhtarov

Differential Revision: D2074044