Use MPMCQueue instead of asox_queue; relaxed notification
Summary: This should help in situations where we have a lot of clients talking to a single
proxy thread (e.g. in HHVM). Other side effects:
- No more wrappers that have to be allocated/deallocated.
- The queue is blocking, so it's important to choose a size that works for our workloads.

This version has a standard notification mechanism (notify an eventfd unless the proxy thread is currently reading), which is still not optimal;
it notifies way too often and also depends on a single atomic state variable (so all threads will contend on it).

Will follow up with another diff with the probabilistic notification idea, which shows significant CPU improvement.

Test Plan: unit tests

Reviewed By: @pavlo-fb

Differential Revision: D2078372