Fix pdstn deadlock
Summary: Fix race codition in ProxyDestinationMap::foreachDestinationSynced that
was triggered when the ProxyDestination was destroyed just after being used
by foreachDestinationSynced, in which case it would try to lock the already
owned mutex.
According to the standard, trying to lock a mutex already owned by the
thread is undefined behavior ("the program may deadlock, or, if the
implementation can detect the deadlock, a resource_deadlock_would_occur error
condition may be thrown").

Test Plan: Unit tests

Reviewed By: @alikhtarov

Differential Revision: D2141237