Workaround for idle CPU usage
Summary: Idle mcrouters still end up calling epoll_wait with a timeout all the time which is apparently expensive
(~15% cpu util with 32 threads).
It's safe to not schedule the timeout if relaxed notification logic is disabled.
Since we don't use relaxed notification in standalone mcrouters, it will take care of the CPU usage when idle.

Note that when not idle, we use epoll_wait with timeout all the time anyway (we schedule a timeout after network sends),
so this is not a real CPU regression.  The issue here is idle workload.

Test Plan: build mcrouter, run locally, observe no CPU usage

Reviewed By: @int

Differential Revision: D2150466