Simplify McrouterClient lifetime logic
Summary: McrouterClient ownership is shared between user and outstanding requests.
shared_ptr is almost enough, except that user's link is "special" - when we let go of it,
we also need to set disconnected_ to true and cancel any queued up requests (with a different callback).

This diff expresses this logic and gets rid of fragile manual reference counting
(note that we had two different counters, one thread safe, the other not; and we had to involve the request queue
to run things on the correct thread - a huge mess).

Test Plan: unit tests pass

Reviewed By: @pavlo-fb

Differential Revision: D2153791