Remove opts.standalone, opts.realtime_disabled, dead code
Summary: Preliminary refactoring for removing StartupLock.

- Simplifies checks in two places: McrouterClient just needs to know whether its safe to bypass the queue;
McrouterInstance just need to check if it has to spawn proxy threads
- Removes realtime threads option.  Turns out it was effectively disabled - we only allowed it for
standalone only which doesn't spawn proxy threads anyway; the only other thread it was used was stats
thread which is not important. If we want to revisit realtime threads (not clear as latency on fabrics is not an issue;
might be a TAO experiment) the right place is to add options to AsyncMcServer.

Test Plan: unit tests

Reviewed By: @pavlo-fb

Differential Revision: D2153865