Use 'pid' instead of 'child_pid' for accuracy and memcached compatability
Summary: The 'child_pid' really is `mcrouter`'s pid (gotten from `getpid(2)`), so while it technically is the is the child of its parent, 'pid' is more accurate and is compatible with memcached.  In addition the 'pid' key is hardcoded in PHP's `Memcached::getStats()`.
Closes #53


Test Plan: author tested, contbuild, flib, unit tests, manual tests

Reviewed By: @pavlo-fb

Differential Revision: D2133576