Deprecate background and managed mode
Summary: Deprecateing background and managed mode so that we can remove
the code in about two releases.
The reasoning behind that is that there are process supervisors
(e.g. runit) that does that job well and there is no need to mantain
code to do the same thing.

Test Plan: Unit tests and manual tests (checked warnings).

Reviewed By: @pavlo-fb

Differential Revision: D2147948