Fix NULL deref in 'retain source ip' mode and add a test for it
Summary: During various refactorings, the `--retain-source-ip` option was broken and enabling it would lead to a NULL dereference as a result of using values from a `std::move`'d object.
This diff fixes the bug and adds an integration test.

Test Plan: - Tested manually by sending a request via a mcrouter instance before and after.
 - Also added integration test that passes.

Reviewed By: @alikhtarov

Differential Revision: D2462361

fb-gh-sync-id: c4a0a702afdc8e25fe8c3f04456a1ad886d4c9f1