Use folly::small_vector instead of std::vector for writing batches of iovecs
Summary: Currently we allocate vector on a heap, which shows up in perf record (I remember even seeing it when Prateek was loadtesting memcached). With this change we're not going to allocate anything when writing less than 64 iovectors.

Test Plan: tests, loadtest OSS mcrouter and ensure that this doesn't show up in perf record.

Reviewed By: alikhtarov

Differential Revision: D2548973

fb-gh-sync-id: 1d7e6bb79880d2fd0f2ffb1954ade650704f0d4c