Don't store shared_ptr with request context
Summary: We don't actually need to hold shared_ptr, because we run on fibers.
Whenever we destroy AsyncMcClientImpl, we will close the socket and fail all requests (i.e. set the replies and post battons).
Though we still need to have the shared_ptr for the 'updateTimeouts' callback.

Test Plan: tests

Reviewed By: pavlo-fb

Differential Revision: D2579570

fb-gh-sync-id: 0fd0f752d95cd3be1fd757649cb38d8860e3239b