Write ip/port to debug fifos and show in mcpiper
Summary: Add a message header to debug fifos protocol. This header is written
before each message sent through mcrouter debug fifos and it contains,
among other information, the ip and port of the server which exchanged
that message with mcrouter.
Mcpiper than read that message header and displays this information to
improve debuggability.
Note: This version of mcpiper is compatible with the older version of
debug fifos' protocol (i.e. the version that doesn't contain message
header).

Test Plan: Unit tests and manual tests

Reviewed By: pavlo-fb

Differential Revision: D2539437

fb-gh-sync-id: fad49fc62bbb579fa7ba0725c474ba90ac184060