Mock the monotonic clock for tests
Summary: Our tests would be more reliable, more effective, and easier to maintain if we had a way to mock the `Clock` instance. This diff is for collecting preliminary feedback on this implementation.

Test Plan: There are some Singleton destruction-order bugs that still need to be worked out before this can be landed.

Reviewed By: mmcduff

Differential Revision: D2582530

fb-gh-sync-id: b3fa784e25d1d5dd5eec074ee4a27d21deff2147