Simplify command line option for configuration
Summary:
The idea is to use 'config' option instead of mutual-exclusive 'config_str' or
'config_file'
If 'config' has 'str:' prefix - we configure from a string; if it has 'file:'
prefix - we read from a file.

Test Plan:
Local testing with `--config 'import(...)`, `--config 'file:...'`, and `--config '<json>'`.  Also updated some of the tests to take the new option and ran

  fbconfig -r mcrouter && fbmake runtests

Reviewed By: pavlo-fb

Differential Revision: D2632420

fb-gh-sync-id: 6ec4a8e0f0f334a28f6c7cd386ec9a763da15559