Add option enable_lease_pairing to failover routes
Summary:
For lease pairing, the lease tokens are going to be stored on a per-route basis.
In order to achieve that, the enable_lease_pairing option is being moved to the
failover routes and a name is required for these routes when lease pairing is enabled.

Test Plan: Unit tests and manual tests

Reviewed By: pavlo-fb

Differential Revision: D2847670

fb-gh-sync-id: 2c2651d818881a1b6da9411961ac6002c837a429