Handle lease pairing in FailoverRoute only
Summary:
Refactor the entire lease pairing logic. Now, everything is handled inside
FailoverRoute handle.
The global option 'enable_lease_pairing' was replaced by a property in
FailoverRoute with the same name.
The main benefit of the new approach is that now all lease-set operations
traverse the routing tree normally. The only think that the lease pairing
does is remember which path it should take in each FailoverRoute it encounters.

Test Plan: Unit tests and manual tests

Reviewed By: pavlo-fb

Differential Revision: D2852985

fb-gh-sync-id: 6bbc7f2059f6111bbf774424f44ba034d26d8a78