Reply with TypedThriftReply on TypedThriftRequest
Summary:As title. In order to not template tao_mc_request_t class, reply logic is moved into callbacks which get set in onRequest.

Implementation of
  McServerRequestContext::reply(McServerRequestContext&& ctx, TypedThriftReply<M>&& reply);
seems not there yet, so I am following the description on the quip doc: https://fb.quip.com/jXQgAiigBKzP

This diff moved McReply.setDestructor into McServerRequestContext::reply. When the implementation for TypedThriftReply is ready, we need to call the destructor in it as well(if this diff is following right approach).

Reviewed By: alikhtarov

Differential Revision: D3120391

fb-gh-sync-id: 503f7ca0116e38180c7de3a6c6b09cd6842271ad
fbshipit-source-id: 503f7ca0116e38180c7de3a6c6b09cd6842271ad