Support fallback routes in mcrouter
Summary:The idea here is we want to support the idea of routing as some other default route. In config now, we fallback to a sort of default routing when a cluster has no real configuration (this is the replacement to logicalization). This adds support in mcrouter. There's two badnesses of this diff
1) It isn't configurable; this isn't really a problem. Lots of sentinels have special meaning and it's OK to not use this feature.
2) It cements the word 'default' for this usecase. Maybe fallback would be better. Default is a good because that's kind of what this is, but 'default route' already has meaning, so this might overload it.

Reviewed By: pavlo-fb

Differential Revision: D3183684

fb-gh-sync-id: e6b3cef8426625447836f49493d81bb95b76a805
fbshipit-source-id: e6b3cef8426625447836f49493d81bb95b76a805