Make shadow-IDs 64-bit
Summary:While looking at logs for a failing box in a shadow tier, I noticed
all the shadow IDs looked similar - in fact, they all begin with AAAAA[A-P],
i.e., zero followed by the first sixteen letters of the alphabet.  In other
words, the high 32 bits are all zero.

This is despite the fact that we have a template for randomInt64 which appears
to be correctly checking the input type and producing a result of the desired
64-bit width.

However, in GCC's implementation of std::mt19937, the type on which the random
number is computed is uint_fast32_t, which is actually defined in stdint.h to
be a unsigned long - tricking us into using the wrong template.

Reviewed By: alikhtarov

Differential Revision: D3203500

fb-gh-sync-id: 9927e682fffc42360e8a454f714fe1540b8f8b96
fbshipit-source-id: 9927e682fffc42360e8a454f714fe1540b8f8b96