Properly clear WriteBuffers
Summary:
We should clear IOBufs as soon as we don't need them. With the current implementation we can hold them alive indefinitely (e.g. if we never use the whole pool of write buffers).
For services that reply with IOBuf that holds a reference to the item, this causes unwanted behavior (e.g. postponing decref).

Reviewed By: jmswen

Differential Revision: D3274983

fb-gh-sync-id: 7f94a6f6ffc56b67e7ae6df3966ce84baaed5847
fbshipit-source-id: 7f94a6f6ffc56b67e7ae6df3966ce84baaed5847