Protocol level compression (client)
Summary:
Integrate caret protocol compression with AsyncMcClient.
When compression is enabled, the client will send two extra fields on every
request, containing the first codec id and the number of codecs supported.

Server then reads that fields and chooses the best codec to use that is
supported by the client. The server than compresses the data and also sends two
extra fields in the reply: the id of the codec used for compressing, and the
uncompressed size.

The client reads the replies received from the server and, if any extra fields
are present, they are assumed to be related to compression, as described above.

Note: This patch contains only the client side.

Reviewed By: pavlo-fb

Differential Revision: D3268807

fbshipit-source-id: 05ca48c57f619dd7e2cf0ec7b99b4c01e21c6408