Ensure CompressionCodecMap is constructed off fibers
Summary:
The ZSTDCompressionCodec constructor can use a lot of stack space, and it may
currently be called while on a fiber (from initializeAsyncMcClient()). P56569479
shows a backtrace of a crashed mcrouter that tried to initialize ZSTDCompressionCodec
on a fiber.

This diff ensures CompressionCodecMap is always called from the main context.

Reviewed By: andreazevedo

Differential Revision: D3715510

fbshipit-source-id: 229a1b5073f036cef961b2698390815bb0551134