Integrate the controller at client side
Summary:
We integrate the controller at AsyncMcServer and set the drop
probability at the caret header. We tested our algorithm on a memcached
box at tst tier. The following figures are the results of the experiments.

1. The time series of queue size when controller is disabled:
https://www.dropbox.com/s/vjso0rhmnx7xjur/memcached_no_control.pdf?dl=0

We can observe the dynamics of the synthesized traffic.

2. The time series of queue size when controller is enabled and the
control target is 3000:
https://www.dropbox.com/s/djvcgfxdm4a0qik/memcached_control.pdf?dl=0

We observe that our algorithm is able to control the queue size below the
target value by varying the drop probability.

3. The time series of gets per second and cpu utilization:
https://www.dropbox.com/s/6fif2ryyvhzow1o/utilization.png?dl=0

The bump on the left corresponds to experiment 1 and the right bump is
experiment 2. The controller controls the queue size by dropping some
requests.

4. The time series of gets per second and cpu utilization when using
the CPU utilization as the signal for the controller.
https://www.dropbox.com/s/npy2yp1yy9n5avo/utilization_cpu.png?dl=0

In this experiement, we set the target as 5%. The bump on the left is the
controlled traffic, whereas the bump on the right is uncontrolled. In this
figure, we observe that the cpu is always under 5% if the controller is
enabled, while in the uncontrolled setting, the cpu is > 6%. Will
rerun the experiment using production traffic and higher cpu target.

Reviewed By: andreazevedo

Differential Revision: D3729587

fbshipit-source-id: 78678b52f2d3b45f1b73b0a9c6c56d4b99761a37