Closing connections after reaching retransmit threshold
Summary:
The main idea here is to close connections once the retransmits per kb cross some threshold. Here's how we do it:
1. We start at min_rxmit_reconnect_threshold. This is the threshold to close connections.
2. A connection may be particularly bad, and to avoid thrashing, every time we close the connection, we increase the threshold to reconnect by a factor of 2 until it hits the ceiling max_rxmit_reconnect_threshold.
3. Since we increase the threshold we also have to bring it back or the connection will not close. To fix this, when the retransmits per kb drop to less than 1/2 times the threshold, we lower it to 1/2 of the threshold until it hits min_rxmit_reconnect_threshold.

As an added security and to prevent the scenario jmswen outlines in the comment below, we also introduce jitters and wait for some cycles since last closure before attempting to close the connection again.

Reviewed By: jmswen

Differential Revision: D3955956

fbshipit-source-id: 69f2781bd1b2e9c22df88f012181380b60d90133