Make it clear that traverse() and RecordingContext are threadsafe
Summary:
It should be legal to call traverse() from non-proxy threads,
but it wasn't clear from the API (we can always access the proxy from the non-const
context and do unsafe actions; RecordingContext constructor itself was not strictly safe).

This is a tiny step towards a much large refactoring, but should be enough for now practically speaking.

Reviewed By: spalamarchuk

Differential Revision: D4154978

fbshipit-source-id: 3135d7b2e87781c55017f71f7f5234066208cb54