Add support for additional route handles
Summary:
Add field `additional_route_handles` to `service` to allow users to easily use
their own custom route handles.

The format for additional_route_handles is the following:
```
service: {
  additional_route_handles: [
    MyRoute @ "path/to/MyRoute.h"
  ]
};
```

Given the above example, a few assumptions are made:
1. `MyRoute` is defined in `"path/to/MyRoute.h"` (or in one of the files it
includes).
2. A factory function
(`RouteHandlePtr makeDuplicateRoute(factory, json);`) is defined in
`path/to/MyRoute.h` (or in one of the files it includes).
3. The factory function is in the same namespace as the idl.

Reviewed By: jmswen

Differential Revision: D4486154

fbshipit-source-id: 52dfb6a8305c07964a5765376f1ce0ce7a6f0faf