MissFailoverRoute: support for "best reply" return
Summary:
The problem: MissFailoverRoute[A, B, C] was always returning the response
from the LAST destination (C) if we have all-misses. If all A, B and C
are healthy, that's perfectly fine. But imagine that C is down. Then,
despite the fact that A and B returned misses, we respond with the error
to the client (because the last C is down). So we have "false negatives",
and the client will see errorous responses is higher than we expect with
no practical reasons.

The solution: we'd better return the "best" response instead of the last
one. I.e. "the best" means "miss is better than fail". I.e. in the above
case when C is dead, the client will see a miss, not an error
(because A and B are healthy).

Reviewed By: jmswen

Differential Revision: D4637155

fbshipit-source-id: 9c584dea8743e73a90013213cf21ed2a5c7e7da1