Call folly::init at beginning of main functions where we can, otherwise call registrationComplete
Summary:
I saw this crashing test in mcrouter-oss on a diff: P57225402. The advised behavior
is to call `folly::init()` at the beginning of all `main()` functions where possible.
(Sometimes it is not possible if a program does uses `optparse()` or `boost::options`;
mcpiper is such an example.)  If `folly::init()` is not safe, call `registrationComplete()`.

Reviewed By: spalamarchuk

Differential Revision: D4754595

fbshipit-source-id: f0f690d61a4e77b1e0c02f1f7bdf6aed1770faca