Improve visitor API for unions
Summary:
This patch introduces a special method for visiting members of a union.
visitUnionMembers() is called for every single member of the union. The visitor
class itself is responsible for checking whether or not the given member of the
union can be accessed (i.e. the union might be set to another field).

Reviewed By: spalamarchuk

Differential Revision: D4933170

fbshipit-source-id: 47b66b6e2fe5b53de3052bcd913ee09b5ad9c353