Add GoAwayMessage
Summary:
This adds a new mechanism to AsyncMcServer to notify clients that the server is about to shutdown and that clients should stop sending new requests.
This is possible only for connections that use caret protocol.
Old clients will silently ignore this message. New clients will ack the server and will stop sending new messages.

Reviewed By: jmswen

Differential Revision: D4941969

fbshipit-source-id: a1b61a4019b786bd34da69ef9981b260b1249136