Add function that converts json to a carbon struct
Summary:
Add a visitor and a convenience function (called convertFromFollyDynamic) to
convert a json to a carbon struct.
The function takes a folly::dynamic and a carbon struct and fills
the struct with the values found in the folly::dynamic.
It returns a list of errors found when populating the struct with the
folly::dynamic.
The convenience works with inline mixins as well as with nested mixins.

Reviewed By: glamtechie

Differential Revision: D4920246

fbshipit-source-id: 7f22c69a8533fed7a4a8f9e1aa5393e8937eb930