Add a simple dynamic client for carbon
Summary:
Added a virtual class that allows sending requets to a carbon server using a
folly::dynamic as input and ouput.
The idea is that we are going to generate a cmd line tool that uses this client
to send requests and receive replies.
We can even add a parser that parses not just json to folly::dynamic, but also
"python constructions".

Reviewed By: glamtechie

Differential Revision: D4944941

fbshipit-source-id: 4904c6858fdc44fca265aba5260d245b1b0504d5