Fix race in TkoTrackerMap
Summary:
Fixes a subtle race in TkoTracker and TkoTrackerMap.

The following would trigger the problem:
  1) TkoTrackerMap::getSuspectServers() is called. We acquire a lock and start
     iterating over all TkoTracker.
  2) During iteration over the map, we successfully lock the weak_ptr to
     TkoTracker (thus getting a valid shared_ptr).
  3) In another thread, we destroy every other shared_ptr to that TkoTracker,
     meaning that now the shared_ptr acquired in step (2) is the last one.
  4) When the shared_ptr acquired in step (2) goes out of scope, we call the
     destructor of TkoTracker.
  5) The TkoTracker destructor calls TkoTrackerMap::removeTracker(), to remove
     itself from TkoTrackerMap.
  6) TkoTrackerMap::removeTracker() tries to acquired the same lock acquired
     in step 1 (which is UB).

The problem described above was triggering an undefined behavior, and it caused
a deadlock a production.

When considering solutions for this problem, we realized that, if we just fixed
the lock problem, we would cause another issue: modifying the map while
iterating over it.

We then opted for a simple solution: add a vector to delay the destruction
of TkoTracker until after the lock is released.

Reviewed By: jmswen

Differential Revision: D5779966

fbshipit-source-id: 1538980cfea9950ccb43f9b269b758eeeb6faa67