Batch iovecs when calling ::sendmsg
Summary:
Right now we use TCP_CORK to signal kernel that we intend to send more. This saves us in terms of packets (and thus the cost of transmitting them), but we still pay the overhead of syscall + extra boilerplate in kernel (fd loockup and some other logic).
With this diff we collect iovecs on stack and call ::sendmsg once per batch or 128 ioves, whichever is smaller. The downside is that we will get only one callback, and thus some requests might get blocked by the write callback.

To avoid the blocking effect I may try adding some capping logic to avoid having a batch over 10K.

Reviewed By: jmswen

Differential Revision: D5763650

fbshipit-source-id: c2abdc8197999ebe6226433f1e52f28ea061891f